+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] <!-- release-date -->
 
+## [0.11.0] - 2023-10-21
+
 ### Added
 
 - Support YouTube video and livestream player URLs as arguments.
@@ -232,7 +234,8 @@ _(none)_
 - Initial implementation.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/LeoniePhiline/showcase-dl/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/LeoniePhiline/showcase-dl/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/LeoniePhiline/showcase-dl/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/LeoniePhiline/showcase-dl/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/LeoniePhiline/showcase-dl/compare/v0.8.1...v0.9.0
 [0.8.1]: https://github.com/LeoniePhiline/showcase-dl/compare/v0.8.0...v0.8.1